small improvements

+++ stelvio/aws/function/function.py
@@ -291,7 +291,7 @@ class Function(Component[FunctionResources], BridgeableComponent):
             return BridgeInvocationResult(
                 success_result=success,
                 error_result=error,
-                process_time_local=int(run_time * 1000),
+                process_time_local=float(run_time * 1000),
                 request_path=event.get("event").get("path", "N/A"),
                 request_method=event.get("event").get("httpMethod", "N/A"),
                 status_code=success.get("statusCode", -1) if success else -1,
